6.自动化测试/requirements.txt
# API automation framework dependencies
pytest>=7.0
requests>=2.28
pytest-html>=3.0
PyYAML>=6.0

# Install: python3 -m venv .venv && source .venv/bin/activate && pip install -r requirements.txt